LPS-151639 Extract environment overriding logic for portal properties to a new util class, prepare for next
